added notifierserver
When using SNS subscribe for http / https protocol, it requires endpoint to be public http or https, however, in many instances, the grpc client and server may be hosted inside VPC or having private IP, therefore SNS subscribe will not be permitted.

To resolve this issue and allow SNS subscribers be called back whenever SNS topic is published, we will use a set of services called Notifier Gateway (public https hosts on ELB), the Notifier Server (one or more public or private containers or ec2) hosts acting as the grpc server responsible to allow grpc clients to connect and allow server to maintain a stream.  Whenever the gateway receives http/s callback from SNS, the gateway will direct the callback to a specific Notifier Server, where upon receipt, the notifier server can relay the message to the grpc client subscribers.